requirements.txt
numpy==1.20.1
Pillow==8.1.1
requests==2.25.1
pymongo==3.11.3
Flask==2.1.0
pytest==6.2.2
mysql-connector-python==8.0.31